chore: moved Wallet definition file out of methods folder
