+++ src/index.d.ts
@@ -1,7 +1,7 @@
 /// <reference path="types/types.d.ts" />
 /// <reference path="types/Account/Account.d.ts" />
-/// <reference path="types/Wallet/methods/Wallet.d.ts" />
+/// <reference path="types/Wallet/Wallet.d.ts" />
 import { Account } from "./types/Account/Account";
-import { Wallet } from "./types/Wallet/methods/Wallet";
+import { Wallet } from "./types/Wallet/Wallet";
 export { Account, Wallet };
 declare module '@dashevo/wallet-lib';
